Make the destination register configurable.
It is possible that a user may want the copied text to be stored in a
different register for some reason.

One such reason could be if they set the "clipboard" setting to
"unnamedplus" for example. In this case, the user may want to change the
destination register to "+".

It is also possible a user may just want it to be copied to a different
register for other reasons.