Use the clipboard setting for the default register.
The `clipboard` setting is now the basis for what register the plugin
saves to. For example, if you set `clipboard` to `unnamed`, it will save
to the `*` register. If it is set to `unnamedplus` it is saved to the
`+` register. Otherwise, it will save it to the default `"` register.

The `g:YankMatches#ClipboardRegister` is still used however if you want
to override the clipboard setting for this plugin.